Added endpoints

+++ server/handlers/api.go
@@ -14,10 +14,14 @@ import (
   "github.com/als9xd/penny/server/models"
 )
 
+/* Threads */
+
 func GetThread(db *sqlx.DB,c *gin.Context) {
   thread := models.Thread{}
   err := db.Get(&thread,`
-    SELECT id,name FROM thread WHERE id=$1;
+    SELECT thread.id,thread.name,thread.creator_id,array_agg(posts.id)
+    FROM thread INNER JOIN posts
+    WHERE thread.id=$1;
   `,c.Param("id"))
   if err == sql.ErrNoRows {
     c.JSON(http.StatusNotFound,gin.H{
@@ -37,7 +41,9 @@ func GetThread(db *sqlx.DB,c *gin.Context) {
 func GetThreads(db *sqlx.DB,c *gin.Context) {
   threads := []models.Thread{}
   err := db.Select(&threads,`
-    SELECT id,name FROM thread;
+    SELECT thread.id,thread.name,thread.creator_id,array_agg(posts.id)
+    FROM thread
+    INNER JOIN posts;
   `)
   if err != nil {
     log.Fatal(err)
@@ -61,7 +67,7 @@ func CreateThread(db *sqlx.DB,c *gin.Context) {
   err = db.Get(&thread,`
     INSERT INTO thread (name)
     VALUES ($1)
-    RETURNING id,name;
+    RETURNING id,name,creator_id;
   `,c.PostForm("name"))
   if err != nil {
     log.Fatal(err)
@@ -85,7 +91,7 @@ func UpdateThread(db *sqlx.DB,c *gin.Context) {
   err = db.Get(&thread,`
     UPDATE thread SET name = $1
     WHERE id = $2
-    RETURNING id,name;
+    RETURNING id,name,creator_id;
   `,c.PostForm("name"),c.Param("id"))
   if err == sql.ErrNoRows {
     c.JSON(http.StatusNotFound,gin.H{
@@ -102,11 +108,10 @@ func UpdateThread(db *sqlx.DB,c *gin.Context) {
   c.JSON(http.StatusOK, gin.H{"data":thread})
 }
 
-
 func DeleteThread(db *sqlx.DB,c *gin.Context) {
   thread := models.Thread{}
   err := db.Get(&thread,`
-    DELETE FROM thread WHERE id = $1 RETURNING id,name;
+    DELETE FROM thread WHERE id = $1 RETURNING id,name,creator_id;
   `,c.Param("id"))
   if err == sql.ErrNoRows {
     c.JSON(http.StatusNotFound,gin.H{
@@ -121,4 +126,61 @@ func DeleteThread(db *sqlx.DB,c *gin.Context) {
     log.Fatal(err)
   }
   c.JSON(http.StatusOK, gin.H{"data":thread})
+}
+
+/* Posts */
+
+func GetPost(db *sqlx.DB,c *gin.Context) {
+  post := models.Post{}
+  err := db.Get(&post,`
+    SELECT * from post WHERE id = $1;
+  `,c.Param("id"))
+  if err == sql.ErrNoRows {
+    c.JSON(http.StatusNotFound,gin.H{
+      "error": gin.H{
+        "code": http.StatusNotFound,
+        "message": fmt.Sprintf("No post with id '%s'",c.Param("id")),
+      },
+    })
+    return
+  }
+  if err != nil {
+    log.Fatal(err)
+  }
+  c.JSON(http.StatusOK, gin.H{"data":post})
+}
+
+func GetPosts(db *sqlx.DB,c *gin.Context) {
+  posts := []models.Post{}
+  err := db.Select(&posts,`
+    SELECT id,author_id,text FROM post;
+  `)
+  if err != nil {
+    log.Fatal(err)
+  }
+  c.JSON(http.StatusOK,gin.H{"data":posts})
+}
+
+func CreatePost(db *sqlx.DB,c *gin.Context) {
+  var b models.Post
+  err := c.ShouldBind(&b)
+  if err != nil {
+    c.JSON(http.StatusBadRequest,gin.H{
+      "error": gin.H{
+        "code": http.StatusBadRequest,
+        "message": err.Error(),
+      },
+    })
+    return
+  }
+  post := models.Post{}
+  err = db.Get(&post,`
+    INSERT INTO post (name)
+    VALUES ($1)
+    RETURNING id,author_id,text;
+  `,c.PostForm("name"))
+  if err != nil {
+    log.Fatal(err)
+  }
+  c.JSON(http.StatusOK, gin.H{"data":post})
 }
